fix for long hrefs obstructing add/remove buttons in multifile select
When hrefs were too long, the selection list in the multiple file select widget were obstructing the add/remove buttons.
The width is now restricted so that the buttons are always showing.
Screen reader users will always be able to read long hrefs but the end of the hrefs will not be visible for sighted users. So a title attribute has been added to the options so you can hover over the hrefs to see the full href.