retrieve browserId using ajax instead of through the template
This ensures that all browser windows get a unique browserId.
Before this, if you duplicated a tab in chromium, both tabs
would have the same browserId. As a consequence, only the first
tab would be notified of a job completion for some reason, this
seems to have fixed that. It also avoids having to use flashing
to pass the browserId to the browser, which helps
towards solving #57.

This commit also makes sure that a browser window showing a
job view only receives file uploads from that specific job.
This was an issue if you tried creating multiple jobs
simultaneously in separate browser windows.